[#4] refactor: commonize integrated test
