+++ src/infra/repositories/users/users.prisma.repository.spec.ts
@@ -3,8 +3,7 @@ import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { UsersPrismaRepository } from './users.prisma.repository'
 import { PrismaModule } from '../../prisma/prisma.module'
 import { PrismaService } from '../../prisma/prisma.service'
-import { GenericContainer } from 'testcontainers'
-import { execSync } from 'child_process'
+import { setUpIntegratedTest } from '../../../shared/integrated.test.setup'
 
 describe('UsersPrismaRepository', () => {
   let repository: UsersPrismaRepository
@@ -12,42 +11,18 @@ describe('UsersPrismaRepository', () => {
 
   // Prepare a TestContainer for setting up a PostgreSQL instance.
   beforeEach(
-    async () => {
-      // Initialize the container
-      console.log('Initialize PG Container ...')
-      const container = await new GenericContainer('postgres:latest')
-        .withExposedPorts(5432)
-        .withEnvironment({
-          POSTGRES_USER: 'user',
-          POSTGRES_PASSWORD: 'password',
-          POSTGRES_DB: 'test',
-        })
-        .start()
+    () =>
+      setUpIntegratedTest(async _prisma => {
+        prisma = _prisma
 
-      // Get the mappedPort from the container and Set the DATABASE_URL
-      const mappedPort = container.getMappedPort(5432)
-      process.env.DATABASE_URL = `postgresql://user:password@localhost:${mappedPort}/test?schema=public`
+        const module: TestingModule = await Test.createTestingModule({
+          imports: [PrismaModule],
+          providers: [UsersPrismaRepository],
+        }).compile()
+        await module.init()
 
-      // Migrate TestContainer and PostgreSQL image
-      execSync('npx prisma migrate dev', {
-        env: {
-          ...process.env,
-          DATABASE_URL: process.env.DATABASE_URL,
-        },
-      })
-
-      prisma = new PrismaService()
-
-      const module: TestingModule = await Test.createTestingModule({
-        imports: [PrismaModule],
-        providers: [UsersPrismaRepository],
-      }).compile()
-      await module.init()
-
-      repository = module.get<UsersPrismaRepository>(UsersPrismaRepository)
-
-      console.log('Initialize PG Container Done')
-    },
+        repository = module.get<UsersPrismaRepository>(UsersPrismaRepository)
+      }),
     1000 * 60 * 3,
   )
